requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
kiteconnect==5.0.1
python-dotenv==1.0.0
APScheduler==3.10.4
gunicorn==21.2.0
pytz==2023.3
requests==2.31.0
pandas==2.1.1
numpy==1.25.2
plotly==5.17.0  # Interactive charting library
# Technical Analysis Dependencies
#ta-lib==0.4.28  # Technical Analysis Library (optional, for advanced indicators)
scikit-learn==1.3.0  # Machine Learning (for future enhancements)
